rustdoc: Use diagnostics for warning about error with including sources
This warning probably almost never happens, but we should still use the
diagnostic infrastructure. My guess is that the error was added back
before rustdoc used the rustc diagnostic infrastructure (it was all
`println!` and `eprintln!` back then!) and since it likely rarely occurs
and this code doesn't change that much, no one thought to transition it
to using diagnostics.

This warning should probably be converted into an error (it's reporting
an error, after all!) but we should at least start by using diagnostics
for it.

The warning looks like this:

    warning: source code was requested to be rendered, but an error occurred: foo
      --> src/test/rustdoc/smart-punct.rs:3:1
       |
    3  | / #![crate_name = "foo"]
    4  | |
    5  | | //! This is the "start" of the 'document'! How'd you know that "it's" the start?
    6  | | //!
    ...  |
    22 | | //! I say "don't smart-punct me -- please!"
    23 | | //! ```
       | |_______^
       |
       = note: skipping rendering of source code

I wasn't sure how to trigger the warning, so to create that message I
temporarily made rustdoc always emit it. That's also why it says "foo"
instead of a real error message.

Note that the span of the diagnostic starts at line 3 because line 1 of
that file is a (non-doc) comment and line 2 is a blank line.